requirements.txt
# Core LLM + Agents SDK
openai>=1.30.0
openai-agents>=0.2.0

# Data & transforms
pandas>=2.0.0
duckdb>=1.0.0
pyarrow>=12.0.0

# DB connectivity
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0

# HTTP & utilities
requests>=2.31.0
pydantic>=2.6.0
rich>=13.7.0

# Orchestration (optional but included)
prefect>=2.18.0

# Demo API server (optional but included)
fastapi>=0.111.0
uvicorn>=0.30.0

python-dotenv>=1.0.1

PyYAML>=6.0.1